we floating

+++ Assets/Scripts/Enemy.cs
@@ -82,7 +82,7 @@ public class Enemy : MonoBehaviour
         _dnaRewarded = _path.dnaRewarded;
     }
     
-    public void TakeDamage(int dmg)
+    public void TakeDamage(float dmg)
     {
         health -= dmg;
         if (health <= 0)
